requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.9.0
Flask==2.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.32
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
platformdirs==3.0.0
protobuf==3.20.3
pycodestyle==2.10.0
pycparser==2.21
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.4
trio==0.22.0
trio-websocket==0.10.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.19.0
Werkzeug==2.2.3
wsproto==1.2.0
